Added suppoort for toyota 747

+++ modules/pcgenerator.py
@@ -96,7 +96,7 @@ class PCGenerator:
 		global machine_config
 
 		self.handler = handler
-                data_dir = os.path.join(os.path.dirname(__file__), '../data/')
+		data_dir = os.path.join(os.path.dirname(__file__), '../data/')
 		with open("{}/{}.json".format(data_dir, machine_id)) as json_config:
 			machine_config = json.loads(json_config.read())
 		if is_blank:
@@ -107,7 +107,7 @@ class PCGenerator:
 			machine_id,
 			len(self.data[0]),
 			len(self.data),
-			machine_config['stitches'] / len(self.data[0]),
+			(int)(machine_config['stitches'] / len(self.data[0])),
 			vert_repeat,
 			is_blank,
 			is_solid_fill)
